Rename FastlaneCore::Slack to Fastlane::Notification::Slack
